+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * Renamed Regex for all step definition attributes to Expression, as it has a cucumber expression or a regular expression (regex) that matches the step text. (Breaking change) (#639)
 * Introduced a new BuildMetadata class to encapsulate CI metadata properties such as ProductName, BuildUrl, BuildNumber, Remote, Revision, Branch, and Tag. These will be used to populate data in Cucumber Messages. (#658)
 * Add xUnit 3 support (#538)
+* Updated Reqnroll project template to add TUnit test framework support and remove EOL .NET versions (6.0, 7.0), added .NET 9.0 support (#701)
+* Removed support for end-of-life .NET frameworks (.NET 6, .NET 7, .NET Core) (#706)
 
 ## Bug fixes: